FIX: fix primic client not being correctly exported to index.js, provide example of processing blog posts on front end (as console logs)
